requirements.txt
Flask==1.0.2
PyMySQL==1.0.2
google-auth==1.31.0
pymongo
dnspython
requests
python-dotenv